juicing up custom error window
