--- src/App/ChaosRecipeEnhancer.UI/Windows/CustomDialog.xaml
@@ -1,73 +1,0 @@
-﻿<Window x:Class="ChaosRecipeEnhancer.UI.Windows.CustomDialog"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Width="500"
-        SizeToContent="Height"
-        ResizeMode="NoResize"
-        Title="Error">
-
-    <Window.Resources>
-        <!-- Define styles for the hyperlinks -->
-        <Style x:Key="HyperlinkStyle" TargetType="Hyperlink">
-            <Setter Property="FontSize" Value="14"/>
-            <Setter Property="FontWeight" Value="Bold"/>
-            <Setter Property="Foreground" Value="DarkBlue"/>
-            <Setter Property="TextDecorations" Value="Underline"/>
-        </Style>
-
-        <!-- Define style for emphasized text -->
-        <Style x:Key="EmphasizedTextStyle" TargetType="Run">
-            <Setter Property="FontSize" Value="14"/>
-            <Setter Property="FontWeight" Value="Bold"/>
-            <Setter Property="Foreground" Value="Black"/>
-        </Style>
-    </Window.Resources>
-
-    <Grid>
-        <StackPanel>
-            <TextBlock x:Name="ReportTextBlock" TextWrapping="Wrap" Foreground="Black" Margin="15 10 15 0" FontWeight="Bold">
-                <!-- The reporting instruction text -->
-            </TextBlock>
-            <Border Background="#FF2B2B2B" Padding="10" Margin="15" CornerRadius="5">
-                <TextBlock x:Name="ExceptionTextBlock" TextWrapping="Wrap" Foreground="#FFD4D4D4" FontFamily="Consolas" Padding="10">
-                    <!-- The exception message will be set programmatically -->
-                </TextBlock>
-            </Border>
-            <Button Content="Copy Error Details" Click="CopyButtonClick" Margin="0 0 0 10" HorizontalAlignment="Center">
-                <Button.Style>
-                    <Style TargetType="Button">
-                        <Setter Property="Background" Value="#FF2B2B2B"/>
-                        <!-- Dark background color -->
-                        <Setter Property="Foreground" Value="#FFCCCCCC"/>
-                        <!-- Light text color -->
-                        <Setter Property="FontWeight" Value="Bold"/>
-                        <Setter Property="BorderBrush" Value="#FF2B2B2B"/>
-                        <!-- Border color -->
-                        <Setter Property="BorderThickness" Value="2"/>
-                        <!-- Border thickness -->
-                        <Setter Property="Padding" Value="10,5"/>
-                        <!-- Padding inside the button -->
-                        <Setter Property="FontSize" Value="14"/>
-                        <Setter Property="Cursor" Value="Hand"/>
-                        <Setter Property="Template">
-                            <Setter.Value>
-                                <ControlTemplate TargetType="Button">
-                                    <!-- Rounded corners -->
-                                    <Border CornerRadius="5"
-                                            Background="{TemplateBinding Background}"
-                                            BorderBrush="{TemplateBinding BorderBrush}"
-                                            BorderThickness="{TemplateBinding BorderThickness}"
-                                            Padding="{TemplateBinding Padding}">
-                                        <!-- Padding moved to Border -->
-                                        <ContentPresenter HorizontalAlignment="Center"
-                                                          VerticalAlignment="Center"/>
-                                    </Border>
-                                </ControlTemplate>
-                            </Setter.Value>
-                        </Setter>
-                    </Style>
-                </Button.Style>
-            </Button>
-        </StackPanel>
-    </Grid>
-</Window>
