+++ src/App/ChaosRecipeEnhancer.UI/Windows/ErrorWindow.xaml
@@ -1,0 +1,111 @@
+﻿<Window x:Class="ChaosRecipeEnhancer.UI.Windows.ErrorWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Width="400"
+        SizeToContent="Height"
+        ResizeMode="NoResize"
+        Title="Error">
+
+    <Window.Resources>
+        <!-- Define styles for the hyperlinks -->
+        <Style x:Key="HyperlinkStyle" TargetType="Hyperlink">
+            <Setter Property="FontSize" Value="14"/>
+            <Setter Property="FontWeight" Value="Bold"/>
+            <Setter Property="Foreground" Value="DarkBlue"/>
+            <Setter Property="TextDecorations" Value="Underline"/>
+        </Style>
+
+        <!-- Define style for emphasized text -->
+        <Style x:Key="EmphasizedTextStyle" TargetType="Run">
+            <Setter Property="FontSize" Value="14"/>
+            <Setter Property="FontWeight" Value="Bold"/>
+            <Setter Property="Foreground" Value="Black"/>
+        </Style>
+
+        <!-- Shared Button Style -->
+        <Style x:Key="ButtonStyle" TargetType="Button">
+            <Setter Property="FontWeight" Value="Bold"/>
+            <Setter Property="BorderThickness" Value="1"/>
+            <Setter Property="Margin" Value="10 0 10 10" />
+            <Setter Property="Padding" Value="10,5"/>
+            <Setter Property="FontSize" Value="14"/>
+            <Setter Property="Cursor" Value="Hand"/>
+            <Setter Property="Template">
+                <Setter.Value>
+                    <ControlTemplate TargetType="Button">
+                        <Border CornerRadius="5"
+                                Background="{TemplateBinding Background}"
+                                BorderBrush="{TemplateBinding BorderBrush}"
+                                BorderThickness="{TemplateBinding BorderThickness}"
+                                Padding="{TemplateBinding Padding}">
+                            <ContentPresenter HorizontalAlignment="Center"
+                                              VerticalAlignment="Center"/>
+                        </Border>
+                    </ControlTemplate>
+                </Setter.Value>
+            </Setter>
+        </Style>
+
+        <!-- Button Style for More Information -->
+        <Style x:Key="MoreInfoButtonStyle" TargetType="Button" BasedOn="{StaticResource ButtonStyle}">
+            <Setter Property="Background" Value="SeaGreen"/>
+            <Setter Property="BorderBrush" Value="CadetBlue"/>
+            <Setter Property="Foreground" Value="White"/>
+        </Style>
+
+        <!-- Button Style for Copy Error Details -->
+        <Style x:Key="CopyButtonStyle" TargetType="Button" BasedOn="{StaticResource ButtonStyle}">
+            <Setter Property="Background" Value="DarkOrange"/>
+            <Setter Property="BorderBrush" Value="Orange"/>
+            <Setter Property="Foreground" Value="White"/>
+        </Style>
+
+        <!-- Style for Expander -->
+        <Style TargetType="Expander">
+            <Setter Property="Foreground" Value="Black"/> <!-- Set your desired color here -->
+            <Setter Property="HeaderTemplate">
+                <Setter.Value>
+                    <DataTemplate>
+                        <TextBlock Text="{Binding}" FontWeight="DemiBold" Foreground="Black"/>
+                    </DataTemplate>
+                </Setter.Value>
+            </Setter>
+        </Style>
+    </Window.Resources>
+
+    <Grid>
+        <StackPanel>
+            <!-- Header with icon and text -->
+            <DockPanel Margin="15 10 15 0" LastChildFill="True">
+                <Image DockPanel.Dock="Left" Source="../Assets/Images/Error.png" Height="24" Width="24" Margin="0,3,5,0" VerticalAlignment="Center"/>
+                <TextBlock DockPanel.Dock="Right" FontWeight="Bold" Foreground="#EB3341" FontSize="24" VerticalAlignment="Center">Error Encountered</TextBlock>
+            </DockPanel>
+            <!-- Expander for additional details -->
+            <Expander Header="Additional Error Details" IsExpanded="False" FontWeight="Bold" Margin="15">
+                <StackPanel>
+                    <!-- Preamble TextBlock (optional visibility) -->
+                    <TextBlock x:Name="PreambleTextBlock" TextWrapping="Wrap" Foreground="Black" FontWeight="DemiBold" Visibility="Visible" Margin="5 10 5 5">
+                        <!-- Preamble content will be set programmatically -->
+                    </TextBlock>
+                    <!-- Exception Details -->
+                    <Border Background="#FF2B2B2B" Padding="10" Margin="0 5 0 0" CornerRadius="5">
+                        <TextBox x:Name="ExceptionTextBox" TextWrapping="Wrap" Foreground="#FFD4D4D4"
+                                 FontFamily="Consolas" Padding="10" BorderThickness="0" Background="Transparent"
+                                 IsReadOnly="True" IsReadOnlyCaretVisible="False">
+                            <!-- The exception message will be set programmatically -->
+                        </TextBox>
+                    </Border>
+                </StackPanel>
+            </Expander>
+            <TextBlock x:Name="ReportTextBlock" TextWrapping="Wrap" Foreground="Black" Margin="15 0 15 15" FontWeight="Bold">
+                <!-- The reporting instruction text -->
+            </TextBlock>
+            <!-- More Information Button -->
+            <Button x:Name="btnMoreInfo" Content="Check Path of Exile Website Status" Click="MoreInfoButtonClick"
+                    Style="{StaticResource MoreInfoButtonStyle}" />
+            <!-- Copy Error Details Button -->
+            <Button x:Name="btnCopyErrorDetails" Content="Copy Error Details" Click="CopyButtonClick"
+                    Style="{StaticResource CopyButtonStyle}" />
+        </StackPanel>
+    </Grid>
+</Window>
